Provide help for the duplicate Tools menu item problem
If reloading an addon that creates a Tools menu item on init, there will
be duplicate items. This has to be handled on a case-by-case basis since
not all addons are initialized the same way or even have menu items.
Tell users how to avoid this if they wish to do so.